Remove a class not used and move some class to criteria package
